+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [PEP440](https://www.python.org/dev/peps/pep-0440/)
 and by implication, [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+- Added a timeout so the client will raise an error if the connection
+  to eBird freezes
+
 ## [3.4.1] - 2025-02-12
 - Disallow lower case location codes